Revert "ArmPlatformPkg/PrePeiCore: Print the firmware version early in boot"
This reverts commit 56035d1c8b255b33b34cd4625c26bc7d38504e30.

Currently, Ampere Platform uses PcdFirmwareVersionString PCD
as a dynamic PCD. It is not supported to use of the dynamic PCD
in SEC phase and not necessary.

PcdFirmwareVersionString is also modified in DXE phase.
This is not sense when we print inconsistent PcdFirmwareVersionString.

So, temporarily revert the commit for the short term.

Signed-off-by: Tinh Nguyen <tinhn@amperecomputing.com>